Move navigation components in separate folder
