+++ frontend/src/components/Layout.tsx
@@ -1,4 +1,4 @@
-import { NavBar } from "@components/NavBar";
+import { NavBar } from "@components/navigation/NavBar";
 import React from "react";
 
 interface LayoutProps {}
