+++ frontend/src/components/SearchBar.tsx
@@ -1,7 +1,7 @@
 import { SearchIcon } from "@chakra-ui/icons";
 import { Divider, Stack } from "@chakra-ui/layout";
 import { HTMLChakraProps, Input, InputGroup, InputLeftElement } from "@chakra-ui/react";
-import { NextChakraLink } from "@components/NextChakraLink";
+import { NextChakraLink } from "@components/navigation/NextChakraLink";
 import { useCoursesQuery } from "@generated/graphql";
 import React, { useState } from "react";
